+++ src/base/LinkInterface.php
@@ -31,7 +31,7 @@ interface LinkInterface extends SavableComponentInterface
     public function getLabel(): string;
     public function getSelectionLabel(): string;
 
-    public function getInputHtml(string $name, Link $currentLink = null): string;
+    public function getInputHtml(string $name, $field, Link $currentLink = null): string;
     public function getSettingsHtml(): string;
 
     public function getType(): string;