+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 1.0.9- 2018-07-16
+## [Unreleased]
 
 ### Changed
 
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 *   Linkit fields are now previewable table colums in table view #32
 *   Fixed issue with the Craft 2 normailize function when a type is not specified
 *   Fixed issue
+
 
 ## 1.0.8 - 2018-04-24